+++ test/compojure/api/resource_test.clj
@@ -94,12 +94,12 @@
 
   (fact "resource without coercion"
     (let [handler (resource
-                    {:get {:parameters {:query-params {(s/optional-key :y) Long
+                    {:coercion (constantly nil)
+                     :get {:parameters {:query-params {(s/optional-key :y) Long
                                                        (s/optional-key :x) Long}}
                            :handler (fn [{{:keys [x y]} :query-params}]
                                       (ok {:x x
-                                           :y y}))}}
-                    {:coercion (constantly nil)})]
+                                           :y y}))}})]
 
       (handler {:request-method :get}) => (has-body {:x nil, :y nil})
       (handler {:request-method :get, :query-params {:x "1"}}) => (has-body {:x "1", :y nil})